Adding indices
I added indices and the queries that I used to test the indices. The testing queries are commented out. Essentially, I compared the duration of certain queries before and after implementing indices. Since the duration was fast after, the indices are working. I'll add the screenshots of the duration comparison onto our project document.